클라이언트에서 컨트롤러 분리
