--- src/main/java/network/ClientContoroller.java
@@ -1,4 +1,0 @@
-package network;
-
-public class ClientContoroller {
-}
